[SYCL-MLIR][NIT] Merge `SYCLHostOps.cpp` into `SYCLOps.cpp`
Having a separate file for `sycl.host.*` member functions is not a
common pattern in other dialects and does not add much in terms of
compilation time.
