+++ mlir-sycl/lib/Dialect/SYCL/IR/CMakeLists.txt
@@ -34,7 +34,6 @@ add_mlir_dialect_library(MLIRSYCLDialect
   MethodUtils.cpp
   SYCLAttributes.cpp
   SYCLDialect.cpp
-  SYCLHostOps.cpp
   SYCLHostTraits.cpp
   SYCLMethodOpInterface.cpp
   SYCLOps.cpp
